+++ docs/changelog.md
@@ -9,6 +9,8 @@ locking variables when using `set`.
 - Better instructions for deploying as a service.
 - Remove the deprecated `Logger` class, see [log](https://github.com/devOpifex/log) package.
 - Pass `host` to free port fetch function.
+- Add hidden option to force change port for upcoming related service.
+- Middleware reworked to use `onHeaders` as provided by httpuv.
 
 ## ambiorix 1.0.2